+++ safe-coloured-text-gen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for safe-coloured-text-gen
 
+## [0.0.0.5] - 2026-04-12
+
+### Added
+
+* `GenValid ChunkStyle` instance
+
 ## [0.0.0.4] - 2026-04-12
 
 ### Added